Revert "Feat/busola/refactor"
