+++ css/factsheet-admin.css
@@ -202,72 +202,6 @@ input[name="gsdp_program_handbook_url"], input[name="gsdp_student_learning_outco
 	color: #0073aa;
 }
 
-	/* Restricted contributor: disable title, permalink, Program Names, Degree Types */
-	body.gsdp-restricted-contributor #titlewrap input#title {
-		pointer-events: none;
-		background-color: #f0f0f0;
-		color: #666;
-		cursor: not-allowed;
-	}
-
-	body.gsdp-restricted-contributor #edit-slug-box,
-	body.gsdp-restricted-contributor #edit-slug-buttons,
-	body.gsdp-restricted-contributor .edit-slug {
-		pointer-events: none;
-		opacity: 0.5;
-	}
-
-	body.gsdp-restricted-contributor #edit-slug-buttons {
-		display: none;
-	}
-
-	body.gsdp-restricted-contributor #gs-program-namediv {
-		pointer-events: none;
-		opacity: 0.5;
-	}
-
-	body.gsdp-restricted-contributor #gs-program-namediv .inside {
-		background-color: #f0f0f0;
-	}
-
-	body.gsdp-restricted-contributor #gs-program-namediv input,
-	body.gsdp-restricted-contributor #gs-program-namediv select,
-	body.gsdp-restricted-contributor #gs-program-namediv button {
-		cursor: not-allowed;
-	}
-
-	body.gsdp-restricted-contributor #tagsdiv-gs-degree-type {
-		pointer-events: none;
-		opacity: 0.5;
-	}
-
-	body.gsdp-restricted-contributor #tagsdiv-gs-degree-type .inside {
-		background-color: #f0f0f0;
-	}
-
-	body.gsdp-restricted-contributor #tagsdiv-gs-degree-type input,
-	body.gsdp-restricted-contributor #tagsdiv-gs-degree-type select,
-	body.gsdp-restricted-contributor #tagsdiv-gs-degree-type button {
-		cursor: not-allowed;
-	}
-
-	/* EAM / Factsheet Team panel disabled for non-admins */
-	body.gsdp-eam-disabled #gsdp-team-members {
-		pointer-events: none;
-		opacity: 0.5;
-	}
-
-	body.gsdp-eam-disabled #gsdp-team-members .inside {
-		background-color: #f0f0f0;
-	}
-
-	body.gsdp-eam-disabled #gsdp-team-members input,
-	body.gsdp-eam-disabled #gsdp-team-members select,
-	body.gsdp-eam-disabled #gsdp-team-members button {
-		cursor: not-allowed;
-	}
-
-
 @media screen and (max-width: 1200px) {
 
 	.factsheet-float input,
